requirements.txt
django==4.1.5
psycopg2==2.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==3.14.0
ipython
ipdb
gunicorn==20.1.0
